requirements.txt
pyautogen[autobuild]
streamlit
yfinance
beautifulsoup4